+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add RouteResolver to resolve platform.sh routes into their environment URLs
 - Resolve routes.yaml route key like domain into their environment domain
+- Read local .platform/routes.yaml if no $PLATFORM_ROUTES is available (upstream routes only)
 
 ## [1.0.2] - 2017-09-19
 ### Changed